Create LICENSE file - GNU GPL v3
Added a file of the license to the mod itself, to include it with all future distributions.  Chose the same license originally selected, as listed on the forum and mod portal: GNU General Public License v3.  Handled this through a pull request to give people more notice; will make the commit in a day or two if no issues arise.